+++ client/src/components/App/Content/InvestigationForm/TabManagement/InteractionsTab/InteractionDialog/InteractionEventForm/ContactsSection/GroupedInvestigations/GroupedInvestigationForm/ContactsForm/ContactsForm.tsx
@@ -1,7 +1,8 @@
 import { Typography } from '@material-ui/core';
 import React from 'react'
 
-import { ConnectedInvestigation } from '../ConnectedInvestigationContact';
+import ConnectedInvestigation from 'models/GroupedInvestigationContacts/ConnectedInvestigation';
+
 import InvestigationAccordion from './InvestigationAccordion/InvestigationAccordion';
 
 const formHeadline = 'מאומתים המקובצים לחקירה :';
