+++ client/src/components/App/Content/InvestigationForm/TabManagement/InteractionsTab/InteractionDialog/InteractionEventForm/ContactsSection/GroupedInvestigations/GroupedInvestigationForm/ContactsForm/InvestigationAccordion/AccordionContent/AccordionContent.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { AccordionDetails } from '@material-ui/core';
+
+import ContactEvent from 'models/GroupedInvestigationContacts/ContactEvent';
+
+interface Props {
+    events : ContactEvent[]
+}
+
+const AccordionContent = (props: Props) => {
+    return (
+        <AccordionDetails>
+            <div>helo</div>
+        </AccordionDetails>
+    )
+}
+
+export default AccordionContent
