+++ client/src/models/GroupedInvestigationContacts/ConnectedInvestigation.ts
@@ -1,0 +1,12 @@
+import ContactEvent from './ContactEvent';
+import CovidPatient from './CovidPatient';
+
+type ConnectedInvestigation = {
+    epidemiologyNumber: number,
+    contactEventsByInvestigationId:{
+        nodes : ContactEvent[]
+    },
+    investigatedPatientByInvestigatedPatientId: CovidPatient
+}
+
+export default ConnectedInvestigation;
