+++ client/src/models/GroupedInvestigationContacts/ConnectedInvestigationContact.ts
@@ -1,0 +1,12 @@
+import ConnectedInvestigation from './ConnectedInvestigation';
+
+type ConnectedInvestigationContact = {
+    investigationGroupReasonByReason: {
+        displayName : string;
+    }
+    investigationsByGroupId: {
+        nodes : ConnectedInvestigation[]
+    }
+}
+
+export default ConnectedInvestigationContact;
