+++ client/src/models/GroupedInvestigationContacts/ContactEvent.ts
@@ -1,0 +1,19 @@
+type ContactEvent = {
+    contactedPeopleByContactEvent : {
+        nodes: {
+            addressByIsolationAddress: {
+                cityByCity: {
+                    displayName: string;
+                }
+            },
+            personByPersonInfo: {
+                firstName : string;
+                lastName : string;
+                id : number;
+                identificationNumber: string;
+            }
+        }[]
+    }
+}
+
+export default ContactEvent;
