+++ client/src/models/GroupedInvestigationContacts/CovidPatient.ts
@@ -1,0 +1,8 @@
+type CovidPatient = {
+    covidPatientByCovidPatient: {
+        fullName: string;
+        identityNumber: string;
+    }
+}
+
+export default CovidPatient;
